strideMCP/pyproject.toml
[project]
name = "stridemcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "fastmcp>=2.11.0",
    "google-genai>=1.30.0",
    "httpx>=0.28.1",
    "matplotlib>=3.10.5",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "sqlalchemy>=2.0.44",
    "stravalib>=2.4",
    "uvicorn>=0.35.0",
]

[tool.uv]
package = true

[project.scripts]
server = "server.main:main"

